+++ CHANGELOG.md
@@ -16,7 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 
-## [0.1.0] - 2024-??-??
+## [0.1.0] - 2024-04-25
 
 ### Added
 - Enhance Documentation